git-svn-id: https://kdlibcpp.svn.codeplex.com/svn@88582 9b283d60-5439-405e-af05-b73fd8c4d996

+++ kdlib/include/kdlib/stack.h
@@ -104,7 +104,7 @@ class Stack {
 public:
 
     unsigned long getFrameCount() {
-        return m_stackTrace.size();
+        return static_cast<unsigned long>( m_stackTrace.size() );
     }
 
     StackFramePtr getFrame( unsigned long frameNumber )
